+++ packages/morpheme_http/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- add fetch download with progress
+- change excludeEndpointUsageToken to List of RegExp
+
 ## 1.0.1
 
 - update documentation and homepage link.